Move TestUtils to top-level package
